test: Engine tests for selection ops
